% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssembleDataForPR.R
\name{CalculatePredictionAndTrueOnDirectInteraction}
\alias{CalculatePredictionAndTrueOnDirectInteraction}
\title{Pair up the direct interaction between different gene pairs and their true co-annotation value}
\usage{
CalculatePredictionAndTrueOnDirectInteraction(data.standard, data.interaction)
}
\arguments{
\item{data.standard:}{the functional standard (Symbol1, Symbol2, 1/0, source (optional))}

\item{data.interaction:}{interaction scores (returned from GetInteractionData)}
}
\value{
A list of :
  data: combined (query) interactions (score) and their corresponding co-annotation (true) (sorted from -ve to +ve)
  Queries.with.AUC: The queries that has some AUC values (others are 0), sorted alphabetically
  auc.neg: AUC value for the data (neg to pos), sorted according to queries
  auc.pos: AUC value for the data (pos to neg), sorted according to queries
}
\description{
Pair up the direct interaction between different gene pairs and their true co-annotation value
}
