% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsembleDataForPR.R
\name{CalculatePredictionAndTrueOnLibraryProfiles}
\alias{CalculatePredictionAndTrueOnLibraryProfiles}
\title{Calculate the profile similarity between all possible pairs, pair up their true co-annotation value and source of the co-annoation.}
\usage{
CalculatePredictionAndTrueOnLibraryProfiles(data.standard, data.interaction)
}
\arguments{
\item{data.standard:}{the functional standard in a format (Gene1 (chr), Gene2 (chr), 1/0, source of annotation (chr))}

\item{data.interaction:}{interaction scores (returned from GetInteractionData)
or pairwise.correlation of interaction data}
}
\value{
a list of vectors
 true: combined.true: true co-ann values (from the standard 0/1) for pairs
 predicted: combined.score: predicted scores (according to similarity) for pairs
 source: origin of the pair (complex, pathway or GO ID)
 indices: exact location of the pair in the co-annoation standard (assumes sorted)
}
\description{
Calculate the profile similarity between all possible pairs, pair up their true co-annotation value and source of the co-annoation.
}
\examples{
 Complex.Pairs <- CalculatePredictionAndTrueOnLibraryProfiles (data.standard, pairwise.correlation)
}
