% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{cTP_func}
\alias{cTP_func}
\title{Plot Category level PR Curve}
\usage{
cTP_func(
  data,
  anno,
  complexes,
  percent_th = 0.05,
  tp_th = 1,
  excludeBG = T,
  out_TP = T
)
}
\arguments{
\item{data}{Stepwise Contribution (numeric) for each complexes}

\item{anno}{Annotation (names) of the complexes in data}

\item{percent_th}{Normalized (by number of pairs) contribution for a complex to be considered as covered}

\item{tp_th}{Minimum number of TPs for a complex to be considered as covered}

\item{out_TP}{}
}
\description{
Plot Category level PR Curve
}
