% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerfCurve.R
\name{GenerateDataForPerfCurve}
\alias{GenerateDataForPerfCurve}
\title{Generate the values for performance curve (Precision and Recall for example) to plot on x and y axis}
\usage{
GenerateDataForPerfCurve(
  value.predicted,
  value.true,
  neg.to.pos = FALSE,
  x.axis = "sensitivity",
  y.axis = "precision"
)
}
\arguments{
\item{value.predicted}{Predicted value / Score (profile similarity / direct interaction)}

\item{value.true}{Actual value (1/0 co-annotation for example)}

\item{neg.to.pos}{assign TRUE to sort from neg to positive (pos -> neg is the default)}

\item{x-axis}{what is the x-axis value you want? (TP, FP, TN, FN, precision, specificity, recall/sensitivity)}

\item{y-axis}{what is the y-axis value you want? (TP, FP, TN, FN, precision, speificity, recall/sensitivity)}
}
\value{
PR -> (x: recall, y: precision), ROC -> (x: recall, y: specificity/fpr)
}
\description{
Generate the values for performance curve (Precision and Recall for example) to plot on x and y axis
}
\examples{
PerfCurve <- function(value.predicted, value.true, neg.to.pos = FALSE, type = 'PR', x.axis = 'sensitivity', y.axis = 'precision')
}
