% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{GetAreaUnderPRCurveForEntities}
\alias{GetAreaUnderPRCurveForEntities}
\title{Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)}
\usage{
GetAreaUnderPRCurveForEntities(summary.standard, data.standard, entity.matrix)
}
\arguments{
\item{summary.standard}{-> Identity of each entity ('ID'), 'Name', Genes insdie ('Gene'), 'Length'}

\item{data.standard}{-> Original Co-annotation data for each entity}

\item{entity.matrix}{-> Output of CalculatePredictionAndTrueOnLibraryProfiles (true, predicted, source, indices)}
}
\value{
Area under the curve for each unique entity (protein complex for example)
}
\description{
Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)
}
