% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{GetCoAnnotationData}
\alias{GetCoAnnotationData}
\title{Reads the Co-annotation standard data}
\usage{
GetCoAnnotationData(filename, delim = "\\t")
}
\arguments{
\item{filename}{Name of the file where the co-annotation data is. Expected: A file with three/four columns: gene1 (string) gene2 (string) co-annotation (0/1) source (IDs).}

\item{delim:}{Delimiter used to separate columns (tab is the default)}
}
\value{
co.annotation The co-annotation data
}
\description{
Reads the Co-annotation standard data
}
