% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{GetContributionOfEntities}
\alias{GetContributionOfEntities}
\title{Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)
Works on the result of CalculatePredictionAndTrueOnLibraryProfiles}
\usage{
GetContributionOfEntities(info.pairs, top.number, data.entity)
}
\arguments{
\item{info.pairs}{-> True, Predicted values with pair association to an entity}

\item{data.entity}{-> Entity ID and Name information}

\item{top.numbers}{-> How much of the signal to consider (from the top)}
}
\value{
The entities and their contribution to the signal
}
\description{
Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)
Works on the result of CalculatePredictionAndTrueOnLibraryProfiles
}
