% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{GetContributionOfEntitiesAtCertainTP}
\alias{GetContributionOfEntitiesAtCertainTP}
\title{Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)
Works on the result of CalculatePredictionAndTrueOnLibraryProfiles}
\usage{
GetContributionOfEntitiesAtCertainTP(
  entity.matrix,
  top.number,
  summary.standard,
  data.standard
)
}
\arguments{
\item{entity.matrix}{-> True, Predicted values with pair association to an entity}

\item{summary.standard}{-> Identity of each entity ('ID'), 'Name', Genes insdie ('Gene'), 'Length'}

\item{data.standard}{-> Co annotation standard (gene1, gene2, 1/0, source)}

\item{top.numbers}{-> How much of the signal to consider (from the top)}
}
\value{
The entities and their contribution to the signal, also the top top.number of TPs
}
\description{
Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)
Works on the result of CalculatePredictionAndTrueOnLibraryProfiles
}
