% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{GetContributionOfEntitiesAtPrecisions}
\alias{GetContributionOfEntitiesAtPrecisions}
\title{Calculate the contribution of complex/PW etc. (num TP) at different precision cutoffs in a tabular format. This is different from Stepwise Contribution in that it doesn't find the smallest subset to explain all the TP, but takes all of them (and therefore some TPs are counted twice i.e. those are in multiple complexes)}
\usage{
GetContributionOfEntitiesAtPrecisions(
  Pairs.in.data,
  list.of.precisions,
  summary.standard
)
}
\arguments{
\item{Pairs.in.data}{Identity of each entity ('ID'), 'Name', Genes insdie ('Gene'), 'Length'}

\item{list.of.precisions}{Precision cutoffs the user is interested in.}

\item{summary.standard}{Summary of the standard used: includes ID, Name, Genes, and Length}
}
\value{
output.complex.contribution -> Complex by precision matrix (where entries are number of TP contributed by the complex at that precision)
This is a entity ID * cutoff data.frame (with an extra column added for entity Name)
}
\description{
Calculate the contribution of complex/PW etc. (num TP) at different precision cutoffs in a tabular format. This is different from Stepwise Contribution in that it doesn't find the smallest subset to explain all the TP, but takes all of them (and therefore some TPs are counted twice i.e. those are in multiple complexes)
}
