% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{GetInteractionData}
\alias{GetInteractionData}
\title{Read the interaction/fitness data given a filename (a usual tab delimited format or R style format)}
\usage{
GetInteractionData(
  filename.input,
  delim = "\\t",
  filename.genes = NULL,
  filename.queries = NULL
)
}
\arguments{
\item{filename.input:}{Input data file; genes on the row and samples on the column. First column provides the gene names, and the rest provide numeric data (interaction, fitness, etc.).}

\item{delim:}{Delimiter used to separate columns (tab is the default)}

\item{filename.genes:}{A file with a set of genes (exclude anything else).}

\item{filename.queries:}{A file with a set of samples (exclude anything else).}
}
\value{
Interaction data (numeric) with genes as rownames() and queries as colnames()
Ensures that the interaction data is sorted alphabetically (by gene names)
}
\description{
Read the interaction/fitness data given a filename (a usual tab delimited format or R style format)
}
\examples{
file.int <- 'Ceres_score_19Q2_with_depmap_id.txt'
data.interaction <- GetInteractionData(file.int)
}
