% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{GetStepwiseContributionOfEntities}
\alias{GetStepwiseContributionOfEntities}
\title{Concentrate the signal (TPs) into the minimum number of complexes that can explain the signal + get the contribution of the complexes at different precisions.}
\usage{
GetStepwiseContributionOfEntities(Pairs.in.data, cutoff.all, summary.standard)
}
\arguments{
\item{Pairs.in.data}{Identity of each entity ('ID'), 'Name', Genes insdie ('Gene'), 'Length'}

\item{cutoff.all}{Precision cutoffs the user is interested in.}

\item{summary.standard}{Summary of the standard used: includes ID, Name, Genes, and Length}
}
\value{
output.stepwise.contribution -> Stepwise contribution matrix
This is a entity ID * cutoff data.frame (with an extra column added for entity Name)
}
\description{
Concentrate the signal (TPs) into the minimum number of complexes that can explain the signal + get the contribution of the complexes at different precisions.
}
