% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessing.R
\name{getSubsetOfCoAnnRemoveGenes}
\alias{getSubsetOfCoAnnRemoveGenes}
\title{Remove a set of genes to re-evaluate the data. Should be used only when we can't use the sister function getSubsetOfCoAnnRemoveIDs or getSubsetOfCoAnnRemovePairs.}
\usage{
getSubsetOfCoAnnRemoveGenes(
  data_standard,
  data_subset,
  gene_list,
  replace = FALSE
)
}
\arguments{
\item{data_standard}{A co-annotation standard (data.frame)}

\item{data_subset}{Output of CalculatePredictionAndTrueOnLibraryProfiles with index column (data.frame)}

\item{gene_list}{a vector of genes to remove. This will, for positive examples, remove all the pairs that include the gene. (character vector)}

\item{replace}{Way1 (replace = false): Remove the positive pairs associated to ids.
This reduces the size of the data a bit.
Way2 (replace = true): Convert the positive pairs to negatives. 
This will maintain the size (number of pairs) of the data.}
}
\description{
Remove a set of genes to re-evaluate the data. Should be used only when we can't use the sister function getSubsetOfCoAnnRemoveIDs or getSubsetOfCoAnnRemovePairs.
}
