% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{GroupUniqueElements}
\alias{GroupUniqueElements}
\title{Calculate the unique elements and their ranges}
\usage{
GroupUniqueElements(data)
}
\arguments{
\item{data}{a sorted vector}
}
\value{
group.indices range of indices of unique elements (with rownames as the unique elements themselves)
}
\description{
Calculate the unique elements and their ranges
}
