% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldStandards.R
\name{MakeCoAnnotationFromGeneSymbols}
\alias{MakeCoAnnotationFromGeneSymbols}
\title{Create a co-annotation standard for gene pairs from given gene memberships into entities (complex, pw etc.).}
\usage{
MakeCoAnnotationFromGeneSymbols(
  data_standard,
  overlap_length = 1,
  subset_str = NULL,
  file_location = NULL
)
}
\arguments{
\item{data_standard}{the functional standard as a data.frame ($ID, $Name, $Genes as columns)}

\item{overlap_length}{Number of common annotations to a complex/pathway to qualify as a co-annoation}

\item{subset_str}{Set this (a vector of strings) if interested in a subset of the data (i.e. 'KEGG' for pathway)}

\item{file_location}{The location of the file (path + filename) to store the co-annotation data. If NULL is given, we still return the co_annotation data.}
}
\value{
a data frame for co-annotations of gene pairs
 gene1: combined.true: true co-ann values (from the standard 0/1) for pairs
 gene2: combined.score: predicted scores (according to similarity) for pairs
 is_annotated: 0/1 value (Are the gene pairs co-annotated to the standard)
 ID: Source(s) of the co-annotation
}
\description{
Create a co-annotation standard for gene pairs from given gene memberships into entities (complex, pw etc.).
}
\examples{
 
}
