% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldStandards.R
\name{MakeCoAnnotationMatrixFromGeneSymbols}
\alias{MakeCoAnnotationMatrixFromGeneSymbols}
\title{Create a co-annotation matrix for gene pairs from given gene memberships into entities (complex, pw etc.).}
\usage{
MakeCoAnnotationMatrixFromGeneSymbols(
  data_standard,
  overlap_length = 1,
  file_location = NULL
)
}
\arguments{
\item{data_standard}{the functional standard as a data.frame ($ID, $Name, $Genes as columns)}

\item{overlap_length}{Number of common annotations to a complex/pathway to qualify as a co-annoation}

\item{file_location}{The location of the file (path + filename) to store the co-annotation data. If NULL is given, we still return the co_annotation data.}
}
\value{
a matrix of gene * gene to show if there is a co-annotation (1) or not (0)
}
\description{
Create a co-annotation matrix for gene pairs from given gene memberships into entities (complex, pw etc.).
}
\examples{
 
}
