% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotCategoryPR}
\alias{PlotCategoryPR}
\title{Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)}
\usage{
PlotCategoryPR(
  data_complex,
  pr.stepwise,
  thresholds = NULL,
  legend.names = NULL,
  legend.ltype = NULL,
  ccol = NULL,
  fig.labs = c("Category TP", "Precision"),
  save.figure = FALSE,
  outfile.name = "test_category_PR",
  outfile.type = "pdf"
)
}
\arguments{
\item{data_complex}{Original input for the co-annotation standard}

\item{pr.stepwise}{Stepwise Contribution (TP) per complex at different precisions}

\item{thresholds}{a vector of two thresholds to filter complexes (1) No of TP, (2) Percent of TP pairs captured.}

\item{legend.names}{legends for different curves}

\item{ccol}{colors for each of the PR curves}

\item{save.figure}{set to TRUE to save the figure (fig.title is used as the name)}

\item{outfile.name}{the name of the output file(figure)}

\item{outfile.type}{type of figure to save - 'pdf' (default) or 'png'}
}
\description{
Separate the contribution of signal into individual entries (a complex, pathway, GO Biological Process etc.)
}
