% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotContributionScatter}
\alias{PlotContributionScatter}
\title{Plot a scatter plot of contribution of all complexes}
\usage{
PlotContributionScatter(
  plot.data,
  length.cutoff = 30,
  AUPRC.cutoff = 0.4,
  fig.title = NULL,
  fig.labs = c("AUPRC", "Size"),
  show.text = FALSE,
  save.figure = FALSE,
  outfile.type = "pdf",
  outfile.name = "test_scatter"
)
}
\arguments{
\item{plot.data}{a list or a data.frame with columns 'Name', 'Length', and 'AUPRC'}

\item{fig.title}{title of the figure}

\item{fig.labs}{labels for x and y axis}

\item{show.text}{set TRUE to show the names of plotted complexes}

\item{outfile.type}{type of figure to save - 'pdf' (default) or 'png'}

\item{outfile.name}{the name of the output file(figure)}

\item{save.figure:}{if TRUE, saves the figure as a pdf (using fig.title as name)}
}
\description{
Plot a scatter plot of contribution of all complexes
}
