% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotPRDirect}
\alias{PlotPRDirect}
\title{Plot performance curves on direct interactions (Negative vs Positive)}
\usage{
PlotPRDirect(
  plot.data,
  type.plot = "log",
  fig.title = NULL,
  fig.labs = c("TP", "Precision"),
  outfile.name = "test_DI",
  outfile.type = "pdf",
  save.figure = FALSE
)
}
\arguments{
\item{plot.data}{a list of two lists names x and y (gives the data on x and y axis)}

\item{type.plot}{can be either semilog (on x axis) (default) or regular}

\item{fig.title}{title of the figure}

\item{fig.labs}{x and y axis labels}

\item{outfile.name}{the name of the output file(figure)}

\item{outfile.type}{type of figure to save - 'pdf' (default) or 'png'}

\item{save.figure}{if TRUE, saves the figure after this run (uses the same name as fig.title)}
}
\description{
Plot performance curves on direct interactions (Negative vs Positive)
}
\examples{

}
