% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotPRSimilarity}
\alias{PlotPRSimilarity}
\title{Plot Performance Curves on Profile Similarities}
\usage{
PlotPRSimilarity(
  pred.ca,
  subsample = FALSE,
  neg.to.pos = FALSE,
  type.plot = "log",
  fig.title = NULL,
  fig.labs = c("TP", "Precision"),
  legend.names = NULL,
  legend.color = c("blue"),
  save.figure = FALSE,
  outfile.name = "test_PR_sim",
  outfile.type = "pdf",
  is.bgdline = FALSE
)
}
\arguments{
\item{pred.ca}{a list of elements each contain two lists predicted (score) and true (co-anotation)}

\item{subsample}{Subsamples from the PR space to reduce the size of the plotted data.}

\item{neg.to.pos}{assign TRUE to sort from neg to positive (pos -> neg is the default)}

\item{type.plot}{can be either 'log' - semilog (x axis) or 'normal' plot}

\item{fig.title}{title of the figure (default nothing)}

\item{fig.labs}{x and y axis labels - default: c('TP', 'Precision')}

\item{legend.names}{names of legends}

\item{legend.color}{colors of legends}

\item{save.figure}{set to TRUE to save the figure in a pdf file (fig.title is used as the name)}

\item{outfile.name}{the name of the output file(figure)}

\item{outfile.type}{type of figure to save - 'pdf' (default) or 'png'}

\item{is.bgdline}{set to TRUE if we want to plot a background (reference) line for precision (y-axis). If plotting multiple plots, setting this to false (default) probably makes more sense unless all of those share the same background.}
}
\value{

}
\description{
Plot Performance Curves on Profile Similarities
}
\examples{

}
