% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{ReformatDepMapData}
\alias{ReformatDepMapData}
\title{Read the DepMap data and convert it to a gene by screen format}
\usage{
ReformatDepMapData(filename, delim = ",")
}
\arguments{
\item{filename}{Name/location of the input depmap file.}

\item{delim:}{Delimiter used to separate columns (',' is the default as it's usually used in depmap files)}
}
\value{
a data.frame in the format of genes (row) * screens (columns)
}
\description{
Read the DepMap data and convert it to a gene by screen format
}
