% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldStandards.R
\name{MakeFuncNetFromGIANT}
\alias{MakeFuncNetFromGIANT}
\title{Download the GIANT functional network and process it to use for FLEX}
\usage{
MakeFuncNetFromGIANT(file_location = NULL)
}
\arguments{
\item{file_location}{The location save the input GIANT file. If it is not there, it will be downloaded.}
}
\value{
a data frame for co-annotations of gene pairs
 gene1: Name of the first gene in the gene pair
 gene2: Name of the second gene in the gene pairs
 is_annotated: 0/1 value (Are the gene pairs functionally related)
}
\description{
Download the GIANT functional network and process it to use for FLEX
}
