% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotContributionStructure}
\alias{PlotContributionStructure}
\title{Plot contribution structure (diversity) of the complexes (a muller plot)}
\usage{
PlotContributionStructure(
  plot.data,
  cutoff.all = NULL,
  min.pairs = 10,
  min.precision.cutoff = 0.5,
  num.complex.to.show = 10,
  show.legend = TRUE,
  list.of.complexes.to.show = NULL,
  alternative.names = NULL,
  ccol = NULL,
  y.lim = NULL,
  fig.title = NULL,
  fig.labs = c("Fraction of TP", "Precision"),
  outfile.name = "test_cont_str",
  outfile.type = "pdf",
  save.figure = FALSE
)
}
\arguments{
\item{plot.data}{a complex (unique) vs precision matrix where each element denote number of TP at that combination}

\item{cutoff.all}{all the precision cutoffs used here (the same as used for plot.data)}

\item{min.pairs}{all the precision cutoffs used here (the same as used for plot.data)}

\item{min.precision.cutoff}{How far down should we go in precision cutoff to calcualte contribution of complexes? Default is 0.5, meaning we calculate contributions starting from the highest precision (where we have at least min.pairs) to the min.precision.cutoff and take a mean contribution to rank the complexes.}

\item{num.complex.to.show}{How many complexes we want to show (everything else will be put to others)?}

\item{show.legend}{default TRUE. Set to FALSE if we don't want to print legends.}

\item{list.of.complexes.to.show}{Use this list of complex to show on the plot, regardless of their ranking.}

\item{alternative.names}{Provide a list of names (corresponds to list.of.complexes.to.show) to use as alternative to original names}

\item{ccol}{colors for the top complexes highlighted (top 10 contributing complexes are colored by default)}

\item{fig.title}{title in case we want to save the image}

\item{fig.labs}{x and y axis labels}

\item{outfile.name}{name of the output file}

\item{outfile.type}{type of figure to save - 'pdf' (default) or 'png'}

\item{save.figure}{do we want to save or just plot}
}
\description{
Plot contribution structure (diversity) of the complexes (a muller plot)
}
